feat(src): move Args struct to args.h
